+++ R/as_named_list.R
@@ -12,6 +12,8 @@
 #' b <- 2
 #' str(as_named_list(a, b))
 #'
+#' as_named_list(a, x = b)
+#'
 #' # an example application for this function is managing saving and
 #' # loading values into the workspace.
 #' if(FALSE) {
@@ -54,22 +56,24 @@ as_named_list <- function(...) {
   # get environment to work in
   unpack_environment <- parent.frame(n = 1)
   # capture ... args
-  name_args <- as.list(do.call(bquote, list(substitute(list(...)), where = unpack_environment)))[-1]
-  n_args <- length(name_args)
+  dot_args <- as.list(do.call(bquote, list(substitute(list(...)), where = unpack_environment)))[-1]
+  n_args <- length(dot_args)
   if(n_args <= 0) {
     stop("wrapr::as_named_list expected arguments")
   }
-  if(length(names(name_args)) > 0) {
-    stop("wrapr::as_named_list did not expect any argument to be bound by name")
-  }
+  arg_names <- names(dot_args)
   str_args <- character(n_args)
   for(i in seq_len(n_args)) {
-    arg_i <- name_args[[i]]
-    if(is.null(arg_i)) {
-      stop("wrapr::as_named_list expected all arguments to not be NULL")
-    }
-    if(!(is.name(arg_i) || is.character(arg_i))) {
-      stop("wrapr::as_named_list expected all arguments to be names or character")
+    if((i <= length(arg_names)) && (nchar(arg_names[[i]]) > 0)) {
+      arg_i <- arg_names[[i]]
+    } else {
+      arg_i <- dot_args[[i]]
+      if(is.null(arg_i)) {
+        stop("wrapr::as_named_list expected all unnamed arguments to not be NULL")
+      }
+      if(!(is.name(arg_i) || is.character(arg_i))) {
+        stop("wrapr::as_named_list expected all unnamed arguments to be names or character")
+      }
     }
     carg_i <- as.character(arg_i)
     if(length(carg_i) != 1) {
@@ -86,9 +90,13 @@ as_named_list <- function(...) {
   if(length(unique(str_args)) != n_args) {
     stop("wrapr::as_named_list, expected all argument names to be unique")
   }
+  # get the values
   values <- vector(mode = 'list', n_args)
   for(i in seq_len(n_args)) {
-    values[[i]] <- get(str_args[[i]], envir = unpack_environment, mode = 'any', inherits = TRUE)
+    vi <- eval(dot_args[[i]], envir = unpack_environment, enclos = unpack_environment)
+    if(!is.null(vi)) { # get arround writing null deletes convention
+      values[[i]] <- vi
+    }
   }
   names(values) <- str_args
   return(values)